requirements-core.txt
fastapi==0.104.1
uvicorn[standard]==0.30.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
psycopg[binary]==3.1.12
pydantic==2.5.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
netmiko==3.4.0
pyats==25.6
genie==25.6
python-dotenv==0.19.0
email-validator==2.1.1
